// src/app/_core/models/peniche-bill-account-vo.ts
import { PenicheBillAccountLineVO } from './peniche-bill-account-line-vo';
import { PenicheTypeEnvoiLivrable } from '../enum/billing-account.enum';

export interface PenicheBillAccountVO {
  identifier: string;
  customerIdentifier: string;
  universe: string;
  alias: string;
  status: string;
  nicheAdmissionDate: string;
  terminationDate: string;
  paymentMode: string;
  paymentCondition: string;
  accountName: string;
  bankCode: string;
  bankName: string;
  officeCode: string;
  officeName: string;
  bankAccountNumber: string;
  ribKey: string;
  payerTitle: string;
  payerLastName: string;
  payerFirstName: string;
  payerCompany: string;
  payerAddress2: string;
  payerAddress3: string;
  payerAddress4: string;
  payerAddress5: string;
  payerAddress6: string;
  payerCedex: string;
  payerDeliveryInfo: string;
  payerOrasId: number;
  payerInseeId: string;
  payerRivoliId: string;
  payerStreetNumber: number;
  payerStreetExtension: string;
  payerStreetType: string;
  payerStreetName: string;
  payerGeocodeX: number;
  payerGeocodeY: number;
  payerSiren: string;
  payerZipCode: string;
  payerCity: string;
  payerCountry: string;
  titularTitle: string;
  titularLastName: string;
  titularFirstName: string;
  titularCompany: string;
  titularAddress4: string;
  titularAddress2: string;
  titularAddress3: string;
  titularAddress5: string;
  titularAddress6: string;
  titularCedex: string;
  titularDeliveryInfo: string;
  titularOrasId: number;
  titularInseeId: string;
  titularRivoliId: string;
  titularStreetNumber: number;
  titularStreetExtension: string;
  titularStreetType: string;
  titularStreetName: string;
  titularGeocodeX: number;
  titularGeocodeY: number;
  titularZipCode: string;
  titularCity: string;
  titularCountry: string;
  titularSiren: string;
  recurrent: boolean;
  balance: number;
  representedAmount: number;
  billingDay: number;
  timeBeforeBilling: number;
  mailNotification: string;
  idMailNotification: number;
  mailOldNotification: string;
  idMailOldNotification: number;
  billAccountLines: PenicheBillAccountLineVO[];
  controlCF: boolean;
  creationDate: string;
  justificationControle: string;
  tvaExemption: boolean;
  typeEnvoi: PenicheTypeEnvoiLivrable;
  compteRenduAttendu: boolean;
  isPluriel: boolean;
  isPro: boolean;
  active: boolean;
  company: boolean;
  particular: boolean;
  payerFullName: string;
  ibanCode: string;
  cbAmexToken: string;
  cbAmexDateFin: string;
}